Scene Clean (#86)
* clean up for big scene test

Added safety check for the checklist ui.
Fixed minor material issue with the rod

* Update RodController.cs

Fixed issue when a fish would be detected from the rod. It spring joint would be removed, but that is what is used to connect the fish to the rod, thus, error. It now adds a spring joint if none is found.